+++ app/Filament/Resources/EmailDeliveryLogResource/Pages/ListEmailDeliveryLogs.php
@@ -1,1 +1,17 @@
-<?php namespace App\Filament\Resources\EmailDeliveryLogResource\Pages; use App\Filament\Resources\EmailDeliveryLogResource; use Filament\Resources\Pages\ListRecords; class ListEmailDeliveryLogs extends ListRecords { protected static string $resource=EmailDeliveryLogResource::class; }
+<?php
+
+namespace App\Filament\Resources\EmailDeliveryLogResource\Pages;
+
+use App\Filament\Resources\EmailDeliveryLogResource;
+use App\Filament\Widgets\EmailDeliveryLogOverview;
+use Filament\Resources\Pages\ListRecords;
+
+class ListEmailDeliveryLogs extends ListRecords
+{
+    protected static string $resource = EmailDeliveryLogResource::class;
+
+    protected function getHeaderWidgets(): array
+    {
+        return [EmailDeliveryLogOverview::class];
+    }
+}